Move worker component into sub-package to break cycle
Having the worker component and the WCI-specific worker in the same go package leads to a dependency cycle when using the worker in OSS. This break apart should fix that cycle.
