+++ wci/workercomponent/fx.go
@@ -1,0 +1,33 @@
+package workercomponent
+
+import (
+	"go.temporal.io/server/common/dynamicconfig"
+	"go.temporal.io/server/common/sdk"
+	workercommon "go.temporal.io/server/service/worker/common"
+	"go.uber.org/fx"
+)
+
+type (
+	componentDeps struct {
+		fx.In
+		ClientFactory sdk.ClientFactory
+	}
+
+	fxResult struct {
+		fx.Out
+		Component workercommon.PerNSWorkerComponent `group:"perNamespaceWorkerComponent"`
+	}
+)
+
+var Module = fx.Options(
+	fx.Provide(NewResult),
+)
+
+func NewResult(
+	dc *dynamicconfig.Collection,
+	params componentDeps,
+) fxResult {
+	return fxResult{
+		Component: NewWCIPerNSWorkerComponent(dc, params.ClientFactory),
+	}
+}
